refactor: Move SettingsProxyModel into alert-common for ProxyManager access
