+++ component/src/main/java/com/synopsys/integration/alert/component/settings/proxy/validator/SettingsProxyValidator.java
@@ -15,8 +15,8 @@ import org.springframework.stereotype.Component;
 
 import com.synopsys.integration.alert.common.descriptor.config.field.errors.AlertFieldStatus;
 import com.synopsys.integration.alert.common.descriptor.config.field.errors.AlertFieldStatusMessages;
+import com.synopsys.integration.alert.common.rest.model.SettingsProxyModel;
 import com.synopsys.integration.alert.common.rest.model.ValidationResponseModel;
-import com.synopsys.integration.alert.component.settings.proxy.model.SettingsProxyModel;
 
 @Component
 public class SettingsProxyValidator {
