+++ component/src/main/java/com/synopsys/integration/alert/component/settings/proxy/web/SettingsProxyController.java
@@ -15,10 +15,10 @@ import com.synopsys.integration.alert.common.rest.AlertRestConstants;
 import com.synopsys.integration.alert.common.rest.ResponseFactory;
 import com.synopsys.integration.alert.common.rest.api.StaticUniqueConfigResourceController;
 import com.synopsys.integration.alert.common.rest.api.ValidateController;
+import com.synopsys.integration.alert.common.rest.model.SettingsProxyModel;
 import com.synopsys.integration.alert.common.rest.model.ValidationResponseModel;
 import com.synopsys.integration.alert.component.settings.proxy.action.SettingsProxyCrudActions;
 import com.synopsys.integration.alert.component.settings.proxy.action.SettingsProxyValidationAction;
-import com.synopsys.integration.alert.component.settings.proxy.model.SettingsProxyModel;
 
 @RestController
 @RequestMapping(AlertRestConstants.SETTINGS_PROXY_PATH)
